requirements.txt
apache-airflow-providers-amazon==8.22.0
pandas==2.2.2
pyarrow==16.1.0
s3fs==2024.5.0
boto3==1.34.69
